chore(health): move to /packages
