+++ packages/administration/src/health/index.ts
@@ -1,0 +1,2 @@
+export { HealthModule } from './health.module';
+export { JhiHealthCheckComponent } from './health.component';
