--- src/app/admin/health/index.ts
@@ -1,1 +1,0 @@
-export { HealthModule } from './health.module';
